requirements.txt
asgiref==3.5.2
autopep8==1.6.0
backports.zoneinfo==0.2.1
Django==4.0.5
django-crispy-forms==1.14.0
django-filter==21.1
gunicorn==20.1.0
pycodestyle==2.8.0
sqlparse==0.4.2
toml==0.10.2
